add JWT-extended and HEROKU configuration requirement

+++ resources/item.py
@@ -1,6 +1,11 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
-import sqlite3
+from flask_jwt_extended import (
+    jwt_required, 
+    get_jwt_claims, 
+    jwt_optional, 
+    get_jwt_identity,
+    fresh_jwt_required
+)
 from models.m_item import ItemModel
 
 
@@ -18,13 +23,14 @@ class Item(Resource):
         help="Every item should have a store id"
     )
 
-    @jwt_required()
+    @jwt_required
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
             return item.json()
         return {'message': 'Item not found'}, 404
 
+    @fresh_jwt_required
     def post(self, name):
         if ItemModel.find_by_name(name):
             return {'message': "An item with name '{}' already exists.".format(name)}, 400
@@ -36,7 +42,12 @@ class Item(Resource):
             return {"message": "An error occurred inserting the item."}, 500
         return item.json(), 201
 
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message': 'Admin required.'}, 401
+
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
@@ -55,7 +66,15 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    
+    @jwt_optional #can give none = user not logged in
     def get(self):
-        print([item.json() for item in ItemModel.query.all()])
-        return {'items': [item.json() for item in ItemModel.query.all()]}
-        # return {'asd':'aedqw'}
+        # get all id of user that is stored in jwt
+        user_id =  get_jwt_identity()
+        items = [item.json() for item in ItemModel.query.all()]
+        if user_id:
+            return {'items': items}, 200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data availableif you logged in'
+        }, 200
